✨ useFetch that manages a data state (data)
This component defines a custom hook called useFetch that manages a data state (data) and allows you to add new data to this state. The addDados function performs a POST request to send new data to a server and updates the local state successfully if the response does not contain errors. In case of failure, the function handles and displays appropriate error messages.